Move ServiceLoader clients up to make their role clearer.
